Refactored analysis modules
