+++ parsley-core/src/ghc/Parsley/Internal/Frontend/Analysis/Flags.hs
@@ -1,0 +1,7 @@
+module Parsley.Internal.Frontend.Analysis.Flags (AnalysisFlags(letBound), emptyFlags) where
+
+newtype AnalysisFlags = AnalysisFlags {
+  letBound :: Bool
+}
+emptyFlags :: AnalysisFlags
+emptyFlags = AnalysisFlags False
