+++ parsley-core/src/ghc/Parsley/Internal/Frontend/Compiler.hs
@@ -19,31 +19,31 @@ into another representation with a code generation function.
 module Parsley.Internal.Frontend.Compiler (compile) where
 
 import Prelude hiding (pred)
-import Data.Dependent.Map                           (DMap)
-import Data.Hashable                                (Hashable, hashWithSalt, hash)
-import Data.HashMap.Strict                          (HashMap)
-import Data.HashSet                                 (HashSet)
-import Data.IORef                                   (IORef, newIORef, readIORef, writeIORef)
-import Data.Kind                                    (Type)
-import Data.Maybe                                   (isJust)
-import Data.Set                                     (Set)
-import Control.Arrow                                (first, second)
-import Control.Monad                                (void, when)
-import Control.Monad.Reader                         (ReaderT, runReaderT, local, ask, MonadReader)
-import GHC.Exts                                     (Int(..), unsafeCoerce#)
-import GHC.Prim                                     (StableName#)
-import GHC.StableName                               (StableName(..), makeStableName, hashStableName, eqStableName)
-import Numeric                                      (showHex)
-import Parsley.Internal.Core.CombinatorAST          (Combinator(..), ScopeRegister(..), Reg(..), Parser(..), traverseCombinator)
-import Parsley.Internal.Core.Identifiers            (IMVar, MVar(..), IΣVar, ΣVar(..), SomeΣVar)
-import Parsley.Internal.Common.Fresh                (HFreshT, newVar, runFreshT)
-import Parsley.Internal.Common.Indexed              (Fix(In), cata, cata', IFunctor(imap), (:+:)(..), (\/), Const1(..))
-import Parsley.Internal.Common.State                (State, get, gets, runState, execState, modify', MonadState)
-import Parsley.Internal.Frontend.Optimiser          (optimise)
-import Parsley.Internal.Frontend.CombinatorAnalyser (analyse, emptyFlags, AnalysisFlags(..))
-import Parsley.Internal.Frontend.Dependencies       (dependencyAnalysis)
-import Parsley.Internal.Trace                       (Trace(trace))
-import System.IO.Unsafe                             (unsafePerformIO)
+import Data.Dependent.Map                              (DMap)
+import Data.Hashable                                   (Hashable, hashWithSalt, hash)
+import Data.HashMap.Strict                             (HashMap)
+import Data.HashSet                                    (HashSet)
+import Data.IORef                                      (IORef, newIORef, readIORef, writeIORef)
+import Data.Kind                                       (Type)
+import Data.Maybe                                      (isJust)
+import Data.Set                                        (Set)
+import Control.Arrow                                   (first, second)
+import Control.Monad                                   (void, when)
+import Control.Monad.Reader                            (ReaderT, runReaderT, local, ask, MonadReader)
+import GHC.Exts                                        (Int(..), unsafeCoerce#)
+import GHC.Prim                                        (StableName#)
+import GHC.StableName                                  (StableName(..), makeStableName, hashStableName, eqStableName)
+import Numeric                                         (showHex)
+import Parsley.Internal.Core.CombinatorAST             (Combinator(..), ScopeRegister(..), Reg(..), Parser(..), traverseCombinator)
+import Parsley.Internal.Core.Identifiers               (IMVar, MVar(..), IΣVar, ΣVar(..), SomeΣVar)
+import Parsley.Internal.Common.Fresh                   (HFreshT, newVar, runFreshT)
+import Parsley.Internal.Common.Indexed                 (Fix(In), cata, cata', IFunctor(imap), (:+:)(..), (\/), Const1(..))
+import Parsley.Internal.Common.State                   (State, get, gets, runState, execState, modify', MonadState)
+import Parsley.Internal.Frontend.Optimiser             (optimise)
+import Parsley.Internal.Frontend.Analysis              (analyse, emptyFlags, letBound)
+import Parsley.Internal.Frontend.Analysis.Dependencies (dependencyAnalysis)
+import Parsley.Internal.Trace                          (Trace(trace))
+import System.IO.Unsafe                                (unsafePerformIO)
 
 import qualified Data.Dependent.Map  as DMap    ((!), empty, insert, mapWithKey, size)
 import qualified Data.HashMap.Strict as HashMap (lookup, insert, empty, insertWith, foldrWithKey, (!))
